Add a partially java compatible shell.
